# docs/swagger.yaml
basePath: /api
definitions:
...
  SuccessResponse:
    properties:
      success:
        description: Only returns true, not an json object
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Valkyrie REST API Specs. This service uses sessions for authentication
  license:
    name: Apache 2.0
  title: Valkyrie API
  version: "1.0"
paths:
...
  /account/{memberId}/friend:
    delete:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Remove Friend
      tags:
      - Friends
    post:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Send Friend Request
      tags:
      - Friends
  /account/{memberId}/friend/accept:
    post:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Accept Friend's Request
      tags:
      - Friends
  /account/{memberId}/friend/cancel:
    post:
      parameters:
      - description: User ID
        in: path
        name: memberId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/SuccessResponse'
      summary: Cancel Friend's Request
      tags:
      - Friends
...
swagger: "2.0"